requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
alembic==1.13.2

# Redis
redis==5.1.0

# AI/ML
ollama==0.3.3
sentence-transformers==3.1.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.2

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.2
pdfplumber==0.11.4

# Code Execution (sandboxed)
RestrictedPython==7.4

# Similarity & Anti-cheat
rapidfuzz==3.9.7

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
httpx==0.27.2
jinja2==3.1.4

datasketch==1.5.9
